implement submit endpoint
